Test fread return value

+++ rand.h
@@ -28,6 +28,6 @@
 
 void init_rand(void);
 uint32_t random32(void);
-void random_buffer(uint8_t *buf, uint32_t len);
+void random_buffer(uint8_t *buf, size_t len);
 
 #endif
